[ELVIS-1379]fix: delete used formule

+++ app/controllers/formules_controller.rb
@@ -121,6 +121,13 @@ class FormulesController < ApplicationController
 
     authorize! :destroy, formule
 
+    if formule.used?
+      render json: {
+        error: "Cette formule est utilisée par une ou plusieurs inscriptions : elle ne peut pas être supprimée."
+      }, status: :unprocessable_entity
+      return
+    end
+
     formule.destroy
     render json: { message: "Formule deleted" }
   end
@@ -169,7 +176,7 @@ class FormulesController < ApplicationController
     pages = query.total_pages
 
     @formules = query.as_json(
-      methods: :archived?,
+      methods: %i[used? archived?],
       include: {
         activities: {
           only: %i[id display_name]
